[SC-13048][SC-15987] ZORDER BY scalability: add a threshold for splitting rewrite of large partition into multiple jobs
This PR aims to alleviate some pain points of using ZOrder in production. Specifically around dealing with large amounts of data.

See these tickets for background motivation:
https://databricks.atlassian.net/browse/SC-13048
https://databricks.atlassian.net/browse/SC-15987

## What changes were proposed in this pull request?
This PR adds one new configuration option for controlling the target size of Z-Cubes.
It changes the default for the delta max file size to be 20% larger than the default for min file size.

This PR reimagines `OptimizeTask` as a trait with two implementations, one for ZOrder and one for Compaction. The `OptimizeTableCommand.execute` method now takes a sequence of `OptimizeTask`s and executes them, 15 at a time. Previously the OptimizeTask was a sequence of files that would be interpreted differently based on whether we are performing a compaction or z-order optimize. In the Z-Order case each task represented an unbounded amount of data, generally of size 100 G, while in the Compaction case each task was 1 gigabyte or less. In both cases we would attempt to perform 400 of these tasks in one spark job, and 15 of these jobs at a time.
This new strategy expands the OptimizeTask to encompass one delta transaction, allowing the size of each batch to be configured separately for z-order and compaction.

## How was this patch tested?
Unit tests to verify that we break lots of files up into multiple tasks.
Unit test to verify that we will not create Z-cubes of size less than `optimize.zorder.mergeStrategy.minCubeSize.threshold` unless we have to.

Please review http://spark.apache.org/contributing.html before opening a pull request.

Authored-by: David Lewis <david.lewis@databricks.com>
Signed-off-by: Mukul Murthy <mukul.murthy@databricks.com>
GitOrigin-RevId: 8a39e20766696126aa7031ca2eb24d6b480914bb