[DELTA-OSS]Remove PGP plugin for OSS Delta
## What changes were proposed in this pull request?

We enabled automatic content signing to simply the release process. The PGP plugin is not needed any more, so just remove it.

I also removed `bintrayReleaseOnPublish in ThisBuild := false`. Turning this off will make the release invisible until someone goes to Bintray and click a button. This was added basically for release testing and it's not needed now.

## How was this patch tested?

Manually pushed a test release to Bintray and confirmed that Bintray signed files for us.

Author: Shixiong Zhu <zsxwing@gmail.com>

#5203 is resolved by zsxwing/remove-pgp-plugin.

GitOrigin-RevId: a2428c2bd35bf119c0310625734c539384b4d0ba