[SC-17993]Fix an issue when re-adding the same file
If deleting a file and re-adding it back, a `Snapshot` may contain both `AddFile` and `RemoveFile` for this path. If this `Snapshot` is used to build a new `Snapshot`, as `input_file_name` for AddFile and `RemoveFile` is `null` and  the order of them is not stable, InMemoryLogReplay may just delete this new added file if its `RemoveFile` is after `AddFile`.

This PR updates InMemoryLogReplay to remove the tombstone when a file is added. This will ensure `InMemoryLogReplay` always output only one `FileAction` (either `AddFile` or `RemoveFile`) for each path.

This PR also removes unused `stateSize` and `hadoopConf` from `InMemoryLogReplay`.

The new unit test.

Author: Shixiong Zhu <zsxwing@gmail.com>

GitOrigin-RevId: 7c3e96c2d344d7c08383019ac2bbfda22f0f6119