[SC-18133][DELTA] Add a `isPartialWriteVisible` interface in LogStore
The writing of checkpoint files doesn't go through log store, so we add an interface `isPartialWriteVisible` to `LogStore` to let the out-of-band writers know whether to use rename or not.

This is a temporary solution - ultimately it would be good to encapsulate this information within log store.

Add a simple end to end test for different log store implementations.

Author: liwensun <liwen.sun@databricks.com>

GitOrigin-RevId: 5bf04c818d702ea24d28bed981f838d164a8d373