[SC-18796]Delta checkpoint should not fail because of FileAlreadyExistsException
## What changes were proposed in this pull request?

When a stage gets retried, a zombie task may still run and write the checkpoint file. This will cause the new tasks in the retried stage fail because of FileAlreadyExistsException on **S3**.

Since the zombie task actually writes the same checkpoint, we can just make the new task successful if the checkpoint file exists.

I also added a TaskFailureListener to delete temp files when a task fails.

## How was this patch tested?

Jenkins

Author: Shixiong Zhu <zsxwing@gmail.com>

#5401 is resolved by zsxwing/SC-18796.

GitOrigin-RevId: a01ed042ab1f59c32b3e2620b2d56600420724a5