[SC-20256]Save one Snapshot when creating DeltaLog
## What changes were proposed in this pull request?

Right now creating DeltaLog needs to create 2 Snapshots when the last checkpoint version is not the same as the latest version. We load the parquet checkpoint to create a `Snapshot`, then call `update` to pick up latest json files after the checkpoint.

We can load the checkpoint and json files together to save one `Snapshot`.

## How was this patch tested?

Jenkins

Author: Shixiong Zhu <zsxwing@gmail.com>

#5865 is resolved by zsxwing/load.

GitOrigin-RevId: 6ab2ce407f613ecf2cbcd5b6136443176470442d