[SC-20413]Snapshot staleness should use the last update timestamp
## What changes were proposed in this pull request?

Right now we use the timestamp of latest commit to check staleness. This is not great since if a table is old and doesn't have recent commits, we will always run `update` to check, which is wasting a list request.

We can remember the timestamp of the last update, and use it to check staleness. Then we can save a list request when a table is not stale.

## How was this patch tested?

Jenkins

Author: Shixiong Zhu <zsxwing@gmail.com>

#5904 is resolved by zsxwing/update-staleness.

GitOrigin-RevId: 8f6939d7d84997ff0ec9490b3ec41db0519d450e