[SC-21941][WARMFIX] Log cleanup should delete the checksum file for version 0
Right now as we are listing from the delta file name from version 0, it will not return the checkpoint file for version 0. This is usually fine since we don't checkpoint version 0. However, technically, we can create a checkpoint for version 0, so it's better to also handle it by using the checkpoint file name to list.

Authored-by: Shixiong Zhu <zsxwing@gmail.com>
Signed-off-by: Shixiong Zhu <zsxwing@gmail.com>
GitOrigin-RevId: 335904ef1738aebd94d0e30684956c675576f266