[SC-20947] Always make the output attributes nullable when writing to Delta
Even though we change the schema as nullable when writing to Delta, the output attributes may remain not-null. In such cases, when writing to Parquet, we still want to keep the attributes as nullable, to avoid potential corruption with Parquet.

This unfortunately bloats certain parquet file sizes, but the changes in the tests suggest that we weren't doing the right thing in the first place.

Regression test

Author: Burak Yavuz <brkyvz@gmail.com>

GitOrigin-RevId: ee2941458fd1ba1ba62e911e8b1987f219dce10e