[DELTA-OSS-EXTERNAL] Specification for the Delta Transaction Protocol
This PR adds a specification for the Delta Transaction Protocol. The goal of this specification is to allow other implementors to build integrations that read and write from a Delta table. This is an early draft that should contain all the information needed to correctly read from a delta table. There are TODOs later in the document, that will be fleshed out in a future PR, on the additional requirements for modifying a table.

Closes delta-io/delta#153

Closes #6388 from zsxwing/filw1ck1.

Authored-by: Michael Armbrust <michael@databricks.com>
Signed-off-by: Shixiong Zhu <zsxwing@gmail.com>
GitOrigin-RevId: c2969b0f07854a665ac2f3aa5a8aa5d0b624eafd