[SC-22791] Issue a specific message about how the schemas mismatch
## What changes were proposed in this pull request?

If a table is created over an existing location that already has a transaction log, then that log determines the schema of the table. If the new creation DDL specifies a different schema, then table creation fails. In the case where the schema only differs in the metadata of a field, we currently print a message that does not show the metadata. It is highly confusing to the user because the error states that the schemas are different, and then prints two seemingly identical schemas.

This fixes that error message by printing a detailed statement of how the two schemas differ:
- Introduces a new method SchemaUtils.detectDifference that returns the first difference found
- Uses that method instead of equality to test for schema compatibility
- Adds various tests for this methos

Also fixes an issue with the existing method isReadCompatible, which previously ignored the nullability of structs in array types. Added a test case to cover that scenario.

For now I added println() statements to the tests so that one can see what the error messages look like in the various cases. I plan to remove these print statements in a follow-up commit once the code is reviewed.

## How was this patch tested?

- unit tests
- manual tests to see the error messages

Please review https://spark.apache.org/contributing.html before opening a pull request.

Author: Andreas Neumann <andreas.neumann@databricks.com>

#6470 is resolved by anew/sc22791.

GitOrigin-RevId: eba98ffe394a1c0af728d4cd924523a411b11f3e