[SC-22738]Describe detail SQL support for OSS Delta
This PR adds `describe detail` SQL support for OSS Delta. The user can use this command to show the metadata of a Delta table.

`DescribeDeltaDetailsCommand.scala` is split to 3 files so that we can have different implementations of `getPathAndTableMetadata`. The codes for the edge `describe detail` should be the same. The OSS Delta needs to resolve `delta./blah/blah` because we don't have DatabricksSessionCatalog in OSS.

Jenkins

Author: Shixiong Zhu <zsxwing@gmail.com>

GitOrigin-RevId: 65707740e474a600ba31f818b0d12dc1967f8558