[SC-25430][DELTA][FIXIT] Avoid unnecessary filesystem checks in Delta Snapshot creation
We do a lot of unnecessary file listing when creating/updating a Delta Snapshot. We have checks around:
1. Is there a _delta_log directory under each parent directory (recursive)
2. File listing to infer file format partitioning (for both JSON and parquet)
3. File existence checks for each file

These all add latency to the state update of a Delta table, and especially (2) has a linear cost as the table grows larger, which is unacceptable.

This PR introduces a DeltaLogFileIndex to use as part of a `LogicalRelation` and `HadoopFSRelation` to skip (1) and (3) and avoid (2), which are all done by DataSource.

Existing tests + Improved numbers under DeltaFileOperationsSuite, and on average 1 second saved on the loading of `prod_ds.usage_logs`.

Author: Burak Yavuz <brkyvz@gmail.com>
Author: Burak Yavuz <burak@databricks.com>

GitOrigin-RevId: 2e25a5b4d4fd821af00dd28942544ba539751d2a