Convert AddFile to FileStatus directly to save FileSystem RPC calls (#10)
This PR creates `FileStatus` directly from `AddFile` to save FileSystem RPC calls. As `AddFile` doesn't have the block locations, we lose the locality. But this is fine today since most of storage systems are on Cloud and the computation is running separately.

I also fixes a bug that we return incorrect file paths to Hive when the partition values have some special values that are escaped in the file path.