[SC-26788][DELTA] Fix issue with DateType partition columns in ConvertToDelta
## What changes were proposed in this pull request?

DateType and TimestampType partitions were not working in ConvertToDelta, because we were not passing in the timezone information during casting. Before we would get a `None.get` error being thrown from the Cast, trying to see if the directory partition value when cast to the data type returned null.

## How was this patch tested?

Regression test

Author: Burak Yavuz <burak@databricks.com>
Author: Burak Yavuz <brkyvz@gmail.com>

#7288 is resolved by brkyvz/convert.

GitOrigin-RevId: 598bd4ee434672cf71e778208a13b24061a055ba