[SC-24886][DELTA] Evaluate partition predicates on the right data type
## What changes were proposed in this pull request?

Right now, we use the wrong data type for partition columns when evaluating partition predicates for the Delta file index - they are always strings. This almost always works because of implicit casting, but there are a few edge cases where implicit casts are not added and we must explicitly set the right type.

SC-10573 was an earlier patch for one such case; we made it minimal to safely warmfix it, but kinda dropped the ball on doing the non-minimal fix.

## How was this patch tested?

new unit test, along with the old one for the old patch

Closes #6743 from jose-torres/fixparttype.

Authored-by: Jose Torres <joseph.torres@databricks.com>
Signed-off-by: Jose Torres <joseph.torres@databricks.com>
GitOrigin-RevId: fc63605f97da3f937cf8f33cda7f7d504a024d36