[SC-24982][DELTA] Do not run jobs on an empty snapshot
When replaying a log for an uncheckpointed Delta table, we first compute an empty snapshot. While this is correct, it does nearly double the time spend on the executor for these tables.

This PR changes this by using the InitialSnapshot class for the initial replay. This class has been modified to make sure it avoids expensive computation (constructing dataframes, caching & executing jobs).

Added UT to `DeltaSuite`.

Author: herman <herman@databricks.com>
Author: Youngbin Kim <youngbin.kim@databricks.com>

GitOrigin-RevId: 5cb7fe0517f3a8f9c032657cc27851ea15da0c62