[SC-28594] Add MetadataGetter interface
This PR introduces the `MetadataGetter` trait, which  provides an interface for getting the metadata information of a Delta log. Also, the `StateMetadataGetter` class is added that implements the `MetadataGetter` interface and extracts the metadata information from the state of a Delta log (`Dataset[SingleAction]`).

It's a refactoring, passes the existing tests.

Author: Ali Afroozeh <ali.afroozeh@databricks.com>

GitOrigin-RevId: dbe250c55567892f4a13d2ab911a5e53f43cad76