[SC-29289][DELTA] Refactor SnapshotManagement and Checksums
## What changes were proposed in this pull request?

This PR refactors the newly introduced SnapshotManagement and MetadataGetter classes, as well as where checksums are verified. We were missing a case, where `getSnapshotAt` was not verifying checksums. The checksum verification has been moved into `MetadataGetter`, since this class is responsible for computing the statistics around a Delta table, and is the thin-waist where we can perform this check.

We also unify some code in SnapshotManagement, and also introduce a method for computing the initial state of a Delta table when it is first loaded. This avoids a two step:
 1. Create initial empty snapshot, because checkpoint doesn't exist
 2. Now do a list and update the table
initialization and consolidates them into one, reducing total file system operations

## How was this patch tested?

Existing tests as this is a pure refactor

Author: Burak Yavuz <brkyvz@gmail.com>

#8112 is resolved by brkyvz/checksumRefactor.

GitOrigin-RevId: 5ad11307c39c5444b2feacb0ad74a9cf89711b1d