[SC-29737][DELTA] Fix tableId's for Delta tables for all tests
## What changes were proposed in this pull request?

We fix the tableIds in all tests to break the assumption that all tableId's must be unique. Users may copy their tables to new locations resulting in duplicate tableIds. None of the logic within Delta code should assume that tables have unique ids.

## How was this patch tested?

A test that required a new table Id in DeltaSourceSuite actually failed, so this change actually works.

Author: Burak Yavuz <brkyvz@gmail.com>
Author: Burak Yavuz <burak@databricks.com>

#8749 is resolved by brkyvz/tableId.

GitOrigin-RevId: 0ef93c1cb18829db315dec7746a7ed6a352e4d3f