[SC-29465][DELTA] Fix Convert To Delta for tables in the HiveMetaStore
Fixes the support for CONVERT TO DELTA on tables stored in the HiveMetaStore. This feature will be open sourced after Spark 3.0 is released.

To ensure that the feature works, I had to refactor some of the test suites, so that the table based tests can run using Hive test utilities as well.

Unit tests

Author: Burak Yavuz <brkyvz@gmail.com>
Author: Burak Yavuz <burak@databricks.com>

Author: Burak Yavuz <brkyvz@gmail.com>
Author: Burak Yavuz <burak@databricks.com>

#8701 is resolved by brkyvz/hiveConvert.

GitOrigin-RevId: fe3057f52c2c85053a69585e7fd7e26b2c240642