[SC-31514][DELTA] Refactor to simplify snapshot computation
This PR removes the re-use of the previous Snapshot's state for the next Snapshot's computation. This logic adds some complexity around having to keep track of RDD lineage to avoid stack overflows, and requiring us to have 2 different Snapshot computation methods when we're trying to load a given version of the table versus just the latest version. Instead we will now perform file listing from the latest known checkpoint at all times and give the full set of files required to build a Snapshot.

The only advantage of reusing the state from the previous snapshot was that it would avoid a cost of hitting the storage system for the existing data. We ran some benchmarks to show that this cost is minimal and should be dwarfed by the time it takes to actually perform ETL.

 - Write a very small commit to a large table (500,000 actions in checkpoint, 106 MB in size) + 8 delta files (100 actions per commit) (This should be a pathological case, because we need to hit the storage system every commit)

Old code path: 1.7 seconds
New code path: 8.1 seconds

 - How long it takes to run a rateStream with 10 rows per second to a Delta table (effects on latency) for 100 batches:

Old code path: 5.3 minutes
New code path: 4.6 minutes
^^ I honestly don't know how it got better. I would've expected it to get slower above. Could be variability in Cloud instance performance

These numbers should be dwarfed by the time to actually write the full data, therefore it seems like a worthy compromise.

Author: Burak Yavuz <brkyvz@gmail.com>

GitOrigin-RevId: afe702686df982766794f498969761c320c42e42