[SC-31282][DELTA] Refactor the composition of Snapshot and Checksum Validation
## What changes were proposed in this pull request?

This PR reorganizes some of the information passed into Snapshot and MetadataGetter to reduce the complexity of the implementations. Checksum validation is something that the Snapshot performs instead of the MetadataGetter now. We also get rid of the MetadataGetter interface, as it doesn't really add any value.

## How was this patch tested?

Existing unit tests

Author: Burak Yavuz <brkyvz@gmail.com>
Author: Burak Yavuz <burak@databricks.com>

#9061 is resolved by brkyvz/refValidation.

GitOrigin-RevId: 437d0ffbcd62c6253422a902d5a6c6e886fbf147