[SC-30821] Fix test flakiness
After looking at failures, I found that the root cause is `enableExpiredLogCleanup` is not set to `false` in some tests and the automatic log cleanup gets triggered. This flag is not flipped because these tests don't have any `Metadata` in the commit logs. `spark.databricks.delta.properties.defaults.enableExpiredLogCleanup` is picked up only if there is a `Metadata` action committed.

This PR adds `startTxnWithManualLogCleanup` and uses it to do the first commit to make sure we disable `enableExpiredLogCleanup` correctly.

Author: Shixiong Zhu <zsxwing@gmail.com>

GitOrigin-RevId: d136dd842b29e165bfba8dfe079094d00e8cb5db