[SC-33262][SS] Don't preserve the original dataframe in the Delta scala API.
## What changes were proposed in this pull request?

In the scala API, when we create a DeltaTable object, we reuse the same dataframe forever. This isn't correct, and can lead to errors if the delta table's schema has changed between DeltaTable initialization and the time of executing an action such as merge. We should instead recreate it from the current state of the delta table when executing actions.

## How was this patch tested?

new unit test

Author: Jose Torres <joseph.torres@databricks.com>

#9382 is resolved by jose-torres/fixmerge.

GitOrigin-RevId: 70c28441197328ad89398c936d0efd95bfa87fd9