[SPARK-31450][SQL] Make ExpressionEncoder usage thread-safe
ExpressionEncoders are not thread-safe. We had various (nasty) bugs because of this.

This PR makes the `ExpressionEncoder` usage of Delta safer and (more) reusable. The function objects are not thread safe, however they are documented as such and should be used in a more limited scope (making it easier to reason about thread safety).

Existing tests.

Author: herman <herman@databricks.com>

GitOrigin-RevId: ac79e712f081fba8c828d4343d1458db425fe9ed