[SC-34318][DELTA] Add test utility class in Python to configure Delta's extensions
## What changes were proposed in this pull request?

Introduces a new utility class DeltaTestCase for Python that sets up a properly configured SparkSession that can be reused across python test suites. In addition, we add the API DeltaTable.forName to check that the suite is in fact being correctly configured.

## How was this patch tested?

By the use of DeltaTable.forName to see that it works.

Author: Burak Yavuz <brkyvz@gmail.com>
Author: Burak Yavuz <burak@databricks.com>

#9801 is resolved by brkyvz/pythonDeltaDDL.

GitOrigin-RevId: 378d155b368bf8db91d4f65b75e41eb92e95c868