[SC-29721][DELTA] Ensure schema changes on a Delta DataFrame throw good errors
Delta has the ability to auto-update for queries. This auto-updating can sometimes be problematic, because a query can change the schema of the source table so drastically that it would be impossible to read the new data files with the schema fixed during analysis. This PR adds a nice error message for cases where this can happen.

Unit tests in DeltaSuite

Author: Burak Yavuz <brkyvz@gmail.com>
Author: Burak Yavuz <burak@databricks.com>

GitOrigin-RevId: 04f2436188bd30b89c8f1cbb09fd483a9812feb9