[ES-24686][DELTA] Fix path conversion in CONVERT TO DELTA.
## What changes were proposed in this pull request?

We currently use `table.location.getPath` to determine the path of a catalog table. But this is a URI path, which is relative to the host, so it won't work for things like s3 buckets where the host is not root. We have to instead do `new Path(table.location).toString`.

## How was this patch tested?

new unit test

Author: Jose Torres <joseph.torres@databricks.com>

#9973 is resolved by jose-torres/fixthething.

GitOrigin-RevId: eec1cb75204b7c147f5e5a7f653c9ca50715aab3