[SC-34720][DELTA] Add check for INSERT INTO an empty directory
## What changes were proposed in this pull request?

DML operations like Update/Delete/Merge require a table to exist. So does INSERT INTO, when you use table names. However, with INSERT INTO, we allow the case for an INSERT when a table directory does contain a `_delta_log` directory, even if it is empty. This adds a check so that we ensure that we are writing into a table that exists, i.e. has at least one commit and will unify the behavior of these DML operations.

## How was this patch tested?

New unit tests

Author: Burak Yavuz <brkyvz@gmail.com>

#9828 is resolved by brkyvz/stopDML.

GitOrigin-RevId: 1416577c510d3fd22b92890299f5ea61785cad52