[DELTA-OSS-EXTERNAL] Fixed doc generation
For java docs, we copy the jquery library file generated in the Scala API docs and inject it into the Java API docs and using it later to show the "evolving" badges. scala docs changed. However, the jquery library was changed (probably due to scala 2.11 to scala 2.12 dependency change) from `jquery.js` to `jquery.min.js` causing the script to fail.

Closes delta-io/delta#449

Signed-off-by: Tathagata Das <tathagata.das1565@gmail.com>

Author: Tathagata Das <tathagata.das1565@gmail.com>
Author: Tathagata Das <tdas@databricks.com>

#10535 is resolved by tdas/cxe28xsf.

GitOrigin-RevId: ac5d1d349fd02f6a7032cbe5d3ea8bbc05a441a7