[SC-37255][DELTA] Get rid of an extra list from the isValid call
## What changes were proposed in this pull request?

Remove the `isValid` check in DeltaLog creation as it is unnecessary. We already call `update()` while running any Delta query or operation. The result of this call can be equivalent to the same thing that `isValid` accomplishes, which is to invalidate the cached snapshot in the case that the underlying table has been recreated.

This change would save us a couple listFrom calls to the underlying storage system, and save users both money and time.

We should be okay after removing this change, because:
 1. update() is called during analysis when running queries
 2. update() is called right before an OptimisticTransaction and we also have a check to ensure that our commit version has not gone backwards after a transaction
 3. Streaming queries always call `listFrom` from the latest version, and don't need isValid in the first place

## How was this patch tested?

New unit tests

Author: Burak Yavuz <brkyvz@gmail.com>
Author: Burak Yavuz <burak@databricks.com>

#10603 is resolved by brkyvz/isValid.

GitOrigin-RevId: f5955e43fb2d2d6d53d9b213c18e4fc7b4f1f71a