[SC-23547] Added Delta Option to use MetadataLog on CONVERT TO DELTA
## What changes were proposed in this pull request?
The use of the parquet MetadataLog during a CONVERT TO DELTA is now a user flag as opposed to the current use of the metadatalog by default.

## How was this patch tested?
A unit test was written to ensure all data is read (not just streamed data in the metadatalog) when this flag is set to false.

Author: kian-db <kian.ghodoussi@databricks.com>

#10802 is resolved by kian-db/convert_option.

GitOrigin-RevId: 4f12d0488adc123bae353b9fcf8e59d3f2fd1add