[DELTA] Support querying last commit version in SparkSession
This introduces a new feature to query the version of the last commit written by the current SparkSession across **any** Delta table.

The API is a new SQLConf field in the `SparkSession`’s `SessionState`: `spark.databricks.delta.lastCommitVersionInSession`, that the user can query from SQL or any supported language. The user can simply query a new SQLConf field to find an up-to-date (optional) String encoding of the version of the last commit written by that `SparkSession`.

If no commits have been made by the spark session, querying the key will return `None`, and after commits, it will return `Some(<last_commit_version_in_session>)`.

How much do we care about atomicity for this operation? Currently, it's possible that the following occurs:
- A pair of commits are being written concurrently (say `commit(0)` and `commit(1)`)
- The ordering of the commit => update SQLConf could be:
 1. `commit(0)`
 2. `commit(1)`
 3. `setSQLConf(0)`
 4. `setSQLConf(1)`
- A read between steps 3 and 4 would yield an incorrect result. In order to mitigate this behavior, locks would likely need to attach the update operation to the commit.

We have moved the `setSQLConf` as close as possible to the code performing the deltaLog write, thereby closing the section where we would produce invalid reads to a negligible size.

Unit tests

Author: Zach Schuermann <zachary.zvs@gmail.com>

GitOrigin-RevId: f816852145480297ea7244368f5c4c02d3b15dd9