[SC-44184][Delta] Fixed incorrect results due to ambiguous attribute resolution in self-merge
## What changes were proposed in this pull request?
Merge operation through Scala/Python APIs can cause data corruption when the source DataFrame was built using the same DeltaTable instance as that used for invoking merge. That is something like this.
```
val source = deltaTable.toDF.filter(...)
deltaTable.as("t").merge(source.as("s"), "t.key = s.key")
```
This is because `deltaTable.toDF` always returns the same DF with resolved attribute refs, and any source DF generated from that will get the same attribute references. Therefore when the merge plan is generated by the `DeltaMergeBuilder`, both the children (`source` and `target`) have the same refs. All expressions in the merge plan may resolve successfully but still create an ambiguous resolved plan, something like this.

```
'DeltaMergeInto ('t.key = 's.key), [Delete [, ]]
:- SubqueryAlias t
:  +- Relation[key#257L,value#258L] parquet
+- SubqueryAlias s
   +- Filter (key#257L = cast(4 as bigint))
      +- Relation[key#257L,value#258L] parquet
```

This can lead to incorrect results because all the expression may bind to the incorrect columns when executing the merge in MergeIntoCommand. For example, expressions to copy unmodified target rows can copy instead pick up values from correspondingly-named columns in source row (i.e., pick up value of source's `key#257L` instead of target's `key#257L`).

The solution chosen in this PR is to rewrite all the attribute references in the target plan duplicate attribute refs are detected between source and target. Here are the details of this when this rewrite occurs.

- Rewrite target instead of source - Rewriting refs is a non-trivial challenge in the general plans as shown [here](https://github.com/apache/spark/blob/master/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/analysis/Analyzer.scala#L1167). So we choose to rewrite the target plan which is guaranteed to be simple contains only a few known plans items.

- Only when duplicate refs are detected - We do not want to always rewrite the references because there may be expressions provided by the user that are pre-resolved on the source and target (i.e., when using `target("colName")` format to generate expressions) and arbitrary reference rewrite will cause unnecessary resolution errors in unambiguous queries in existing workloads.

When the duplicated refs in target plan are rewritten, if there are pre-resolved merge expressions (conditions and actions) where the same refs are used, then those are likely to be ambiguous. Hence we unresolve those refs and let the standard resolution logic check whether they are ambiguous or not. **This is the part that there is a slight chance of regression - some existing self-merge with pre-resolved expressions used to accidentally produce the correct results will now start failing. However, the chance of such regression is very low.** More specifically, here is the precise characterization of what queries are affected.

- Non-self-merge queries - Not affected as this PR rewrites plans only when duplicate refs.
- Self-merge queries - there are two cases:
  - Without pre-resolved exprs - There are unambiguous self-merge queries that would generate ambiguous plans therefore most likely produce incorrect results. With this PR, those unambiguous queries will always produce unambiguous plans and therefore correct results. With this PR, those unambiguous queries will always produce unambiguous plans and therefore correct results. No risk of failing those workloads, just making them correct.
  - With pre-resolved exprs - The small set of queries that produced ambiguous plans will now throw resolution error. A very small subset of these queries that passed earlier (but now fails) could have accidentally produced correct results. This the tiny risk.

## How was this patch tested?

- Added more tests covering the presence of pre-resolved refs in all merge expressions
- Added tests for self merge, without and with pre-resolved refs.

Author: Tathagata Das <tathagata.das1565@gmail.com>

#11610 is resolved by tdas/SC-44184.

GitOrigin-RevId: 6809a073f3342b83d6591281f40565ad513f49f3