[SC-44268][DELTA] Update Delta Protocol Doc with Checkpoint format information
## What changes were proposed in this pull request?

In this PR we add sections to the Delta Protocol doc around the checkpoint format, essentially describing what format the checkpoint should be in, how the data should be distributed if this is a multi-part checkpoint, and what the schema of the checkpoint file should be. We also note the new columns that must exist in the new protocol, when we upgrade to writer version 3.

I also noticed that the action reconciliation section was missing, therefore added that as well.

## How was this patch tested?

N/A

Author: Burak Yavuz <brkyvz@gmail.com>

#11673 is resolved by brkyvz/protDoc.

GitOrigin-RevId: e6a4bb90875a1a09b24b123eb928993bee838556