[SC-44271][DELTA] Introduce default protocol version for Delta tables
## What changes were proposed in this pull request?

In this PR, we introduce the concept of a default protocol version for Delta tables. As we build new features for Delta, we may have to make protocol breaking changes. Protocol upgrades could be a high-friction process as all readers and/or writers to a Delta table may have to be upgraded after a protocol change.

With the new default configurations, users will be able to create tables that can be accessed the most widely. Only if they decide to use a new feature, only then will the table protocol require an upgrade. If the table is created with the respective table configuration, then we will choose the minimum required protocol version.

During this process, I found some other irregularities, specifically around table creation. REPLACE wasn't really "creating" a new table as it should in the sense that:
 - Nullability information would be dropped (tests for this will be added separately)
 - default configurations for a table through SQL configs were not being picked up
Therefore I introduced a new `updateMetadataForNewTable` method that will set the protocol for the new table as well. I've also moved protocol verification checks into `updateMetadata` for fail-fast behavior and `prepareCommit` as a last check.

We also remove the "Upgrade your protocol" message as it can become annoying (it was initially meant to be).

## How was this patch tested?

Tests in DeltaProtocolVersionSuite

Author: Burak Yavuz <brkyvz@gmail.com>

#11723 is resolved by brkyvz/protocolConf.

GitOrigin-RevId: df07ae9b8c183f4c9732632d8a7cfcf70522f038