[SC-45188][DELTA] Update the new checkpoint format for Delta
## What changes were proposed in this pull request?

Implement a new format for checkpoints in Delta that contain partition values as parsed columns as part of the checkpoint. This will help in avoiding potential casting bugs when performing partition filters. This new code path will be enabled by default for tables that are upgraded to writer protocol 3. The related code for upgrading the protocol will be introduced in a follow up PR.

## How was this patch tested?

Unit tests

Author: Burak Yavuz <brkyvz@gmail.com>
Author: Burak Yavuz <burak@databricks.com>

#11864 is resolved by brkyvz/checkpointV2New.

GitOrigin-RevId: ad74d3e3b15efc58ce432e56bd197db4411f6c7b