[SC-48540][WARMFIX][DELTA] Introduce table properties writeStatsAsJson and writeStatsAsStruct for the checkpoint format
We introduce the table properties `delta.checkpoint.writeStatsAsJson` and `delta.checkpoint.writeStatsAsStruct` to decide what to include as part of the checkpoint data. We used to consider the protocol version as a requirement for writing the new checkpoint columns. However, it'll be a better design to have these table properties instead and have the protocol upgrade a way to enforce the selection of these table properties.

We also introduce a SQL conf for users to test whether they would want to opt-in to the new format or not, instead of having to make a table property change that can cause transaction conflicts.

Author: Burak Yavuz <brkyvz@gmail.com>

GitOrigin-RevId: 0d7873424ce5aa7c351e3f9a2ccead47118cef37