[SC-49224][DELTA] Fix bug preventing INSERT INTO on nested struct literals
## What changes were proposed in this pull request?

In DeltaAnalysis, we try to construct the appropriate casts for by-position resolution, but we cast nested struct fields to their own names on the source rather than the appropriate name in the target. This produces an infinite loop where DeltaAnalysis detects a cast is needed, inserts another project, but makes no progress.

## How was this patch tested?

new unit test

Author: Jose Torres <joseph.torres@databricks.com>

#12597 is resolved by jose-torres/testbug.

GitOrigin-RevId: b27586d381c01245a62e67c96aca8282d599ac5d